[FLINK-20455][build] Move notice checking into separate class
